requirements.txt
transformers
torch
Pillow
pytest
timm
black
bandit